pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "agent-bench"
version = "0.1.0"
description = "Benchmark AI coding agents against each other on the same task"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [{name = "Hiren Thakore"}]
keywords = ["ai", "benchmark", "coding", "agents", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.scripts]
agent-bench = "agent_bench.cli:cli"

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-mock>=3.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
